+++ src/some/simple/items/Brick.java
@@ -5,7 +5,7 @@ import some.Shape;
 
 public class Brick extends Item {
 
-    public Brick(String name, int weight, int size, String color) {
+    public Brick(String name, double weight, int size, String color) {
         super(name, Shape.RECTANGLE, weight, size, color);
     }
 }
